Added simple unit in fusion matmul optimization
