+++ crates/burn-cubecl-fusion/src/matmul/tune.rs
@@ -32,6 +32,7 @@ pub fn fused_matmul_autotune<R: Runtime, BT: CubeElement>(
     let tunables = TunableSet::new(create_key::<R>, input_gen::<R>)
         .with_tunable(tune_fallback::<R, BT>) // First one should always work.
         .with_tunable(tune_simple_fused::<R, BT>)
+        .with_tunable(tune_simple_unit_fused::<R, BT>)
         .with_tunable_optional(tune_double_buffering_fused::<R, BT>, |key| {
             should_tune_double_buffering(key.num_out_buffers > 1, &key.matmul_key)
         });
@@ -96,6 +97,21 @@ fn tune_simple_fused<R: Runtime, BT: CubeElement>(
         TuneContext::Original(context) => optimization.execute_simple_fused::<BT>(context),
         TuneContext::Fork(mut context_owned) => {
             optimization.execute_simple_fused::<BT>(&mut context_owned.as_context())
+        }
+    }
+    .map_err(|e| format!("{e:?}"))
+}
+
+fn tune_simple_unit_fused<R: Runtime, BT: CubeElement>(
+    input: TuneInput<R, MatmulOptimization<R>>,
+) -> Result<TuneOutput<R>, String> {
+    let optimization = input.optimization();
+    let context = input.context();
+
+    match context {
+        TuneContext::Original(context) => optimization.execute_simple_unit_fused::<BT>(context),
+        TuneContext::Fork(mut context_owned) => {
+            optimization.execute_simple_unit_fused::<BT>(&mut context_owned.as_context())
         }
     }
     .map_err(|e| format!("{e:?}"))
