+++ crates/burn-cubecl-fusion/src/shared/trace/plan.rs
@@ -7,7 +7,7 @@ use crate::{
 use burn_ir::{TensorId, TensorIr};
 use cubecl::Runtime;
 
-use super::block::FuseBlock;
+use super::{block::FuseBlock, vectorization::Vect};
 
 /// The plan is responsible to keep runtime information related to the launch of a fused kernel
 /// at one place.
@@ -80,32 +80,6 @@ impl ReferenceSelection {
         match self {
             ReferenceSelection::Concrete { strides, .. } => strides == strides_inplace,
             _ => false,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-pub enum Vect {
-    Broadcasted,
-    Aligned(u8),
-}
-
-impl Vect {
-    pub fn line_size(&self) -> u8 {
-        match self {
-            Vect::Broadcasted => 1,
-            Vect::Aligned(val) => *val,
-        }
-    }
-
-    pub fn is_broadcast(&self) -> bool {
-        matches!(self, Vect::Broadcasted)
-    }
-
-    pub fn limit_to_one(&self) -> Self {
-        match self {
-            Vect::Broadcasted => Vect::Broadcasted,
-            Vect::Aligned(_) => Vect::Aligned(1),
         }
     }
 }
