+++ crates/burn-cubecl-fusion/src/shared/trace/vectorization/mod.rs
@@ -1,0 +1,5 @@
+mod base;
+mod planner;
+
+pub use base::*;
+pub use planner::*;
